floats and negs for exps in pow

+++ arithmetic.py
@@ -17,10 +17,10 @@ def multiply(arg_list):
     return num1 * num2
 
 def divide(arg_list):
-    num1 = int(arg_list[0])
+    num1 = float(arg_list[0])
     num2 = int(arg_list[1])
 
-    return float(num1)/num2
+    return num1/num2
 
 def square(arg_list):
     num1 = int(arg_list[0])
@@ -34,7 +34,7 @@ def cube(arg_list):
 
 def power(arg_list):
     base = int(arg_list[0])
-    exponent = int(arg_list[1])
+    exponent = float(arg_list[1])
 
     return base ** exponent
